pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlopsproj"
version = "0.0.1"
description = "nicolai"
authors = [{ name = "nicolai", email = "your@email.com" }]

keywords = ["machine learning", "MLOps"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[gs]>=3.66.1",
    "dvc-s3>=3.3.0",
    "fastapi==0.115.6",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "transformers>=4.46.0",
    "safetensors>=0.4.5",
    "pillow>=10.4.0",
    "torchvision>=0.21.0",
    "pytorch-lightning>=2.6.0",
    "tensorboard>=2.20.0",
    "wandb>=0.24.0",
    "hydra-core>=1.3.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "google-cloud-aiplatform>=1.134.0",
    "streamlit>=1.53.1",
    "python-multipart>=0.0.6",
    "markdown>=3.10",
    "loguru>=0.7.3",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "pytest-asyncio>=0.23.0",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
]


[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests as async (using pytest-asyncio)",
]
